cvelate: Move 'cve' types and code into module
